feat: 사진 선택 시 지도 마운트 문제 최적화
