+++ components/circle/album/SinglePhotoIcon.jsx
@@ -8,11 +8,26 @@ import { useRecoilState } from 'recoil';
 import { Image } from 'expo-image';
 import { s3BaseUrl } from '../../../constants/config';
 import { useMutation } from '@tanstack/react-query';
+import { selectedPhotosState } from '../../../stores/circle-store';
 
-export const SinglePhotoIcon = ({ photo, handleSelectedPhotos, selectedPhotos }) => {
+export const SinglePhotoIcon = ({ photo }) => {
   const [circleSelectButtonActive, setCircleSelectButtonActive] = useRecoilState(circleSelectButtonState);
+  const [selectedPhotos, setSelectedPhotos] = useRecoilState(selectedPhotosState);
 
   const navigation = useNavigation();
+
+  // 사진 선택하기
+  const handleSelectedPhotos = photoId => {
+    setSelectedPhotos(prevSelectedPhotos => {
+      if (prevSelectedPhotos.includes(photoId)) {
+        // 이미 선택된 사진이면 제거
+        return prevSelectedPhotos.filter(id => id !== photoId);
+      } else {
+        // 선택되지 않은 사진이면 추가
+        return [...prevSelectedPhotos, photoId];
+      }
+    });
+  };
 
   const navigateToPhotoCom = () => {
     navigation.navigate('PhotoCom', { photo });
