+++ screens/MyPage/Circle/SingleCircle.jsx
@@ -12,13 +12,13 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { deletePhoto, fetchPhotos, fetchSortedPhotos } from '../../../api/photoApi';
 import { useNavigation } from 'expo-router';
 import CircleHeader from '../../../components/header/CircleHeader';
+import { selectedPhotosState } from '../../../stores/circle-store';
 
 export const SingleCircle = ({ route }) => {
   const navigation = useNavigation();
   const { circleId } = route.params; //써클의 id값 찾아내기
   const [isReady, setIsReady] = useState(splashState);
   const [isMap, setIsMap] = useState(true);
-  const [selectedPhotos, setSelectedPhotos] = useState([]);
   const queryClient = useQueryClient();
 
   // 써클의 사진들을 불러오기
@@ -31,15 +31,6 @@ export const SingleCircle = ({ route }) => {
     queryFn: () => fetchPhotos(circleId),
   });
 
-  // 사진 삭제하기
-  const photoDeleteMutation = useMutation({
-    mutationFn: () => deletePhoto(selectedPhotos),
-    onSuccess: data => {
-      queryClient.invalidateQueries('photo');
-      setSelectedPhotos([]);
-    },
-  });
-
   // 사진 정렬하기
   const photoSortMutation = useMutation({
     mutationFn: args => fetchSortedPhotos(args.circleId, args.sortType),
@@ -48,34 +39,6 @@ export const SingleCircle = ({ route }) => {
       // console.log('실시간 정렬');
     },
   });
-
-  // 사진 선택하기
-  const handleSelectedPhotos = photoId => {
-    setSelectedPhotos(prevSelectedPhotos => {
-      if (prevSelectedPhotos.includes(photoId)) {
-        // 이미 선택된 사진이면 제거
-        return prevSelectedPhotos.filter(id => id !== photoId);
-      } else {
-        // 선택되지 않은 사진이면 추가
-        return [...prevSelectedPhotos, photoId];
-      }
-    });
-  };
-
-  // 모든 사진 선택하기
-  const selectAllPhotos = () => {
-    setSelectedPhotos(prevPhotos => {
-      // Check if all photos are already selected
-      const allSelected = prevPhotos.length === photoData.length;
-      // If all photos are selected, deselect all. Else, select all.
-      return allSelected ? [] : photoData.map(photo => photo.id);
-    });
-  };
-
-  // selectedPhotos 비우기
-  const clearSelectedPhotos = () => {
-    setSelectedPhotos([]);
-  };
 
   const handleScroll = e => {
     //스크롤 위치를 확인
@@ -90,7 +53,7 @@ export const SingleCircle = ({ route }) => {
   // 써클의 사진의 id를 전달받아서 사진을 불러오기
   const renderItem = ({ item, index }) => (
     <Pressable style={{ flex: 0.33 }}>
-      <SinglePhotoIcon photo={item} handleSelectedPhotos={handleSelectedPhotos} selectedPhotos={selectedPhotos} />
+      <SinglePhotoIcon photo={item} />
     </Pressable>
   );
 
@@ -109,15 +72,7 @@ export const SingleCircle = ({ route }) => {
           data={photoData}
           numColumns={3}
           keyExtractor={item => item.id}
-          ListHeaderComponent={() => (
-            <HeaderComponent
-              circleId={circleId}
-              photoDeleteMutation={photoDeleteMutation}
-              selectedPhotos={selectedPhotos}
-              selectAllPhotos={selectAllPhotos}
-              clearSelectedPhotos={clearSelectedPhotos}
-            />
-          )}
+          ListHeaderComponent={() => <HeaderComponent circleId={circleId} />}
           onScroll={handleScroll}
           scrollEventThrottle={16}
           renderItem={renderItem}
@@ -129,9 +84,34 @@ export const SingleCircle = ({ route }) => {
   }
 };
 
-const HeaderComponent = React.memo(({ circleId, photoDeleteMutation, selectAllPhotos, clearSelectedPhotos }) => {
-  const [isMap, setIsMap] = useState(true);
+const HeaderComponent = React.memo(({ circleId }) => {
   const [circleSelectButtonActive, setCircleSelectButtonActive] = useRecoilState(circleSelectButtonState);
+  const [selectedPhotos, setSelectedPhotos] = useRecoilState(selectedPhotosState);
+  const queryClient = useQueryClient();
+
+  // 사진 삭제하기
+  const photoDeleteMutation = useMutation({
+    mutationFn: () => deletePhoto(selectedPhotos),
+    onSuccess: data => {
+      queryClient.invalidateQueries('photo');
+      setSelectedPhotos([]);
+    },
+  });
+
+  // 모든 사진 선택하기
+  const selectAllPhotos = () => {
+    setSelectedPhotos(prevPhotos => {
+      // Check if all photos are already selected
+      const allSelected = prevPhotos.length === photoData.length;
+      // If all photos are selected, deselect all. Else, select all.
+      return allSelected ? [] : photoData.map(photo => photo.id);
+    });
+  };
+
+  // selectedPhotos 비우기
+  const clearSelectedPhotos = () => {
+    setSelectedPhotos([]);
+  };
 
   const changeSelection = () => {
     setCircleSelectButtonActive(!circleSelectButtonActive);
@@ -178,7 +158,9 @@ const HeaderComponent = React.memo(({ circleId, photoDeleteMutation, selectAllPh
       <View style={styles.personBox}>
         <OthersProfile circleId={circleId} />
       </View>
-      <View style={styles.mapContainer}>{isMap && <MemorizedSingleMap />}</View>
+      <View style={styles.mapContainer}>
+        <MemorizedSingleMap />
+      </View>
       <View style={styles.wrapper}>
         <Text style={styles.imageText}>사진</Text>
         {!circleSelectButtonActive ? (
